CMS: Extract testcase code to seperate file
